Add needed changes to build on OBS
All single packages lists need all the packages.

Also add keyword to tell the OBS source server to drop the project
signing public certificate in the source directory, so that
mkosi can pick it up and use it. It needs to be in the
mkosi.conf that gets extracted at source runtime.